+++ src/graderPrivate/java/h04/H3_3_Test.java
@@ -2,16 +2,16 @@ package h04;
 
 import fopbot.*;
 import h04.util.H04TestBase;
+import h04.util.TestConstants;
 import h04.util.reflect.StudentLinks;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.Timeout;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.mockito.Mockito;
 import org.mockito.stubbing.Answer;
-import org.objectweb.asm.*;
 import org.sourcegrade.jagr.api.rubric.TestForSubmission;
-import org.sourcegrade.jagr.api.testing.ClassTransformer;
 import org.sourcegrade.jagr.api.testing.TestCycle;
 import org.sourcegrade.jagr.api.testing.extension.JagrExecutionCondition;
 import org.sourcegrade.jagr.api.testing.extension.TestCycleResolver;
@@ -24,11 +24,17 @@ import org.tudalgo.algoutils.tutor.general.reflections.BasicMethodLink;
 import org.tudalgo.algoutils.tutor.general.reflections.Modifier;
 
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.tudalgo.algoutils.tutor.general.assertions.Assertions2.*;
 
 @TestForSubmission
+@Timeout(
+    value = TestConstants.TEST_TIMEOUT_IN_SECONDS,
+    unit = TimeUnit.SECONDS,
+    threadMode = Timeout.ThreadMode.SEPARATE_THREAD
+)
 public class H3_3_Test extends H04TestBase {
 
     public static Object KEY_PRESS_LISTENER = null;
@@ -121,51 +127,10 @@ public class H3_3_Test extends H04TestBase {
     @ExtendWith(TestCycleResolver.class)
     public void testConstructor(TestCycle testCycle) throws Throwable {
         var link = StudentLinks.KEYBOARD_FIELD_SELECTOR_LINK.get();
-        Class<?> clazz = testCycle.getClassLoader().loadClass(link.reflection().getName(), new ClassTransformer() {
-            @Override
-            public String getName() {
-                return "KeyboardFieldSelectorTransformer";
-            }
-
-            @Override
-            public void transform(ClassReader reader, ClassWriter writer) {
-                reader.accept(new ClassVisitor(Opcodes.ASM9, writer) {
-                    @Override
-                    public MethodVisitor visitMethod(int access, String name, String descriptor, String signature, String[] exceptions) {
-                        if (name.equals("<init>") &&
-                            descriptor.equals("()V")) {
-                            return new MethodVisitor(Opcodes.ASM9, super.visitMethod(access, name, descriptor, signature, exceptions)) {
-                                @Override
-                                public void visitMethodInsn(int opcode, String owner, String name, String descriptor, boolean isInterface) {
-                                    if (opcode == Opcodes.INVOKESTATIC &&
-                                        owner.equals("fopbot/World") &&
-                                        name.equals("addKeyPressListener") &&
-                                        descriptor.equals("(Lfopbot/KeyPressListener;)V")) {
-                                        super.visitFieldInsn(Opcodes.PUTSTATIC,
-                                            "h04/H3_3_Test",
-                                            "KEY_PRESS_LISTENER",
-                                            "Ljava/lang/Object;");
-                                        super.visitInsn(Opcodes.ICONST_1);
-                                        super.visitFieldInsn(Opcodes.PUTSTATIC,
-                                            "h04/H3_3_Test",
-                                            "CALLED_ADD_KEY_PRESS_LISTENER",
-                                            "Z");
-                                    } else {
-                                        super.visitMethodInsn(opcode, owner, name, descriptor, isInterface);
-                                    }
-                                }
-                            };
-                        } else {
-                            return super.visitMethod(access, name, descriptor, signature, exceptions);
-                        }
-                    }
-                }, ClassReader.SKIP_DEBUG | ClassReader.SKIP_FRAMES);
-            }
-        });
         var constructor = link.getConstructor(Matcher.of(constructorLink -> constructorLink.typeList().isEmpty()));
         Assertions3.assertCorrectModifiers(constructor, Modifier.PUBLIC);
 
-        var instance = clazz.getDeclaredConstructor().newInstance();
+        var instance = constructor.invoke();
         assertTrue(CALLED_ADD_KEY_PRESS_LISTENER, emptyContext(), result ->
             "World.addKeyPressListener(...) was not called in the constructor");
         assertSame(instance, KEY_PRESS_LISTENER, emptyContext(), result ->
