+++ src/graderPrivate/java/h04/util/TestConstants.java
@@ -9,7 +9,7 @@ public class TestConstants {
     public static final boolean SHOW_WORLD = IS_DEBUGGING;
     public static final int WORLD_DELAY = 500;
 
-    public static final int TEST_TIMEOUT_IN_SECONDS = 2;
+    public static final int TEST_TIMEOUT_IN_SECONDS = 10;
 
     public static final boolean SKIP_AFTER_FIRST_FAILED_TEST = true;
 }
